#3287: Made NERSwitch its own component, removed redundant check
